requirements.txt
aiogram==3.26.0
APScheduler==3.11.2
python-dotenv==1.2.2
